// src/lib/api-types.d.ts
export interface components {
    schemas: {
// ...
        AccountResponseDto: {
            /**
             * @description Hesabın benzersiz ID numarası
             * @example uuid-string
             */
            id: string;
            /**
             * @description Hesabın adı
             * @example Garanti Maaş Hesabı
             */
            name: string;
            /**
             * @description Hesap Tipi (Nakit, Banka, Kredi Kartı vb.)
             * @example BANK
             * @enum {string}
             */
            type: "BANK" | "CREDIT_CARD" | "CASH" | "PERSON";
            /**
             * @description Hesaptaki güncel bakiye
             * @example 15500.5
             */
            balance: number;
            /**
             * @description Para Birimi (TRY, USD, EUR, GOLD)
             * @example TRY
             */
            currency: string;
            /**
             * @description Kredi kartları için hesap kesim günü (Ayın kaçıncı günü)
             * @example 15
             */
            statementDay?: number;
            /**
             * @description Kredi kartları için son ödeme günü (Ayın kaçıncı günü)
             * @example 25
             */
            dueDay?: number;
            /**
             * Format: date-time
             * @description Hesabın sisteme eklenme tarihi
             * @example 2026-03-18T12:00:00Z
             */
            createdAt: string;
            /**
             * Format: date-time
             * @description Hesap bilgilerinin son güncellenme tarihi
             * @example 2026-03-18T15:30:00Z
             */
            updatedAt: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}